Move classes to main file
